requirements.txt
python-barcode
Pillow
opencv-python
pyzbar